requirements.txt
gitpython
pyinstaller